% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSAT.R
\name{MSAT}
\alias{MSAT}
\title{MSAT}
\usage{
MSAT(
  num_items = 18L,
  with_welcome = TRUE,
  with_finish = TRUE,
  label = "MSAT",
  feedback = MSAT_feedback_with_score(),
  dict = MSAT::MSAT_dict
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed. Defaults to  TRUE}

\item{with_finish}{(Scalar boolean) Indicates, if a finish (not final!) page shall be displayed. Defaults to  TRUE}

\item{label}{(Character scalar) Label to give the MSAT results in the output file.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test.}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This function defines a MSAT  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MSAT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
}
\details{
For demoing the MSAT, consider using \code{\link{MSAT_demo}()}.
For a standalone implementation of the MSAT,
consider using \code{\link{MSAT_standalone}()}.
}
