% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{MSAT_feedback_with_score}
\alias{MSAT_feedback_with_score}
\title{MSAT feedback (with score)}
\usage{
MSAT_feedback_with_score(dict = MSAT::MSAT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual feedback at the end of the test.
}
\examples{
\dontrun{
MSAT_demo(feedback = MSAT_feedback_with_score())}
}
