% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MSAT_standalone}
\alias{MSAT_standalone}
\title{Standalone MSAT}
\usage{
MSAT_standalone(
  title = NULL,
  num_items = 18L,
  with_id = TRUE,
  with_feedback = FALSE,
  with_welcome = TRUE,
  admin_password = "test123",
  researcher_email = "robin.hake@uni-oldenburg.de",
  languages = c("en", "de"),
  dict = MSAT::MSAT_dict,
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{num_items}{(Scalar integer) Number of items to be adminstered.}

\item{with_id}{(Scalar boolean) Indicates, if ID should be asked for. Defaults to TRUE}

\item{with_feedback}{(Scalar boolean) Indicates if performance feedback will be given at the end of the test. Defaults to  FALSE}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed.  Defaults to  TRUE}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"en"}),
German (\code{"de"}),  Russian (\code{"ru"}), and Nederlands (\code{"nl"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto" for default validation
which means ID should consist only of  alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{MSAT}()}.}
}
\description{
This function launches a standalone testing session for the MSAT.
This can be used for data collection, either in the laboratory or online.
}
